Location overview & script fix
This includes a new widget: the LocationOverview. It is used to display
a list of all locations in the city the user might visit. To display
some content, a new asset folder is introduced with metadata contained
in yaml-files.

Furthermore, the scripts for ESense sensor recording have been tweaked a
little bit.