Centralize state in `LocationInfoDB`
Previously, LocationInfoDB was only used as a sort of 'loader' that
starts loading the info from the files and sqlite DB and caches them.
The widgets (mainly the map screen and the location overview) read the
list of `LocationInfo` and put them into their own state as well.

This, of course, was bad: each time they wanted to update a LocationInfo
entry they needed to update their own state and also tell the DB what
changed. But even with this doubled state management, the other widgets
were not notified of the change.

Now, all widgets access the cached LocationInfo directly from the
`LocationInfoDB`. If they want to change something, they use according
methods of that class. In order to be able to tell when they need to
update their state, a Stream of updates can be subscribed, not unsimilar
to the Observer pattern.