Add ESense Interact debugpage, new settings layout
The ESense Walking detection is now based on a single value (the Y
        axis).
At the point of inquiry, the class determines if the Y gyroscope is
under a certain value, and if so, returns true.

The autoplay with walking detection can now be disabled in the settings.

The GPS accuracy has been changed from NAVIGATION to HIGH.